requirements.txt
fastapi==0.128.8
uvicorn==0.34.0
python-dotenv==1.2.1
pandas==2.3.3
groq==0.25.0
requests==2.32.4
itsdangerous==2.2.0
starlette==0.46.2
python-multipart==0.0.20
openpyxl>=3.1.0
